+++ doc/source/changelog.rst
@@ -21,6 +21,7 @@ master: 1.8.1
 * Stricter handling of bool type in overload with int types
 * Fix uint64_t template instantiation regression
 * Do not filter out enum data for ``__dir__``
+* Fix lookup of interpreter-only explicit instantiations
 * Further hiding of upstream code/dependencies
 * Extended documentation